requirements.txt
# Runtime (API + Alembic)
alembic>=1.13
asyncpg>=0.29
bcrypt>=4.2
email-validator>=2
fastapi>=0.115
pydantic>=2
pydantic-settings>=2
PyJWT>=2.8
sqlalchemy[asyncio]>=2.0
uvicorn[standard]>=0.30
